requirements.txt
discord.py
steam
Pillow
google-genai
python-dotenv
